Add AnswerAttachmentsRepository interface and AnswerAttachment entity; fix import paths
