+++ src/domain/forum/application/repositories/question-answer-repository.ts
@@ -1,0 +1,6 @@
+import { AnswerAttachment } from "../../enterprise/entities/answer-attachment";
+
+export interface AnswerAttachmentsRepository {
+  findManyByAnswerId(answerId: string): Promise<AnswerAttachment[]>;
+  deleteManyByAnswerId(answerId: string): Promise<void>;
+}
