+++ src/domain/forum/application/repositories/question-attachments-repository.ts
@@ -1,4 +1,4 @@
-import { QuestionAttachment } from "../../enterprise/entities/answer-attatchment";
+import { QuestionAttachment } from "../../enterprise/entities/question-attachment";
 
 export interface QuestionAttachmentsRepository {
   findManyByQuestionId(questionId: string): Promise<QuestionAttachment[]>;
