+++ src/domain/forum/application/use-cases/create-question.ts
@@ -2,7 +2,7 @@ import { UniqueEntityID } from "@/core/entities/unique-entity-id";
 import { Question } from "../../enterprise/entities/question";
 import { QuestionsRepository } from "../repositories/question-repository";
 import { Either, right } from "@/core/either";
-import { QuestionAttachment } from "../../enterprise/entities/answer-attatchment";
+import { QuestionAttachment } from "../../enterprise/entities/answer-attachment";
 import { QuestionAttachmentList } from "../../enterprise/entities/question-attachment-list";
 
 interface CreateQuestionUseCaseRequest {
