+++ src/domain/forum/application/use-cases/edit-question.ts
@@ -6,7 +6,7 @@ import { NotAllowedError } from "./errors/not-allowed-error";
 import { QuestionAttachmentsRepository } from "../repositories/question-attachments-repository";
 import { QuestionAttachmentList } from "../../enterprise/entities/question-attachment-list";
 import { UniqueEntityID } from "@/core/entities/unique-entity-id";
-import { QuestionAttachment } from "../../enterprise/entities/answer-attatchment";
+import { QuestionAttachment } from "../../enterprise/entities/answer-attachment";
 
 interface EditQuestionUseCaseRequest {
   authorId: string;
