+++ src/domain/forum/enterprise/entities/answer-attachment.ts
@@ -1,0 +1,24 @@
+import { Entity } from "@/core/entities/entity";
+import { UniqueEntityID } from "@/core/entities/unique-entity-id";
+
+interface AnswerAttachmentProps {
+  answerId: UniqueEntityID;
+  attachmentId: UniqueEntityID;
+}
+
+export class AnswerAttachment extends Entity<AnswerAttachmentProps> {
+  get answerId() {
+    return this.props.answerId;
+  }
+
+  get attachmentId() {
+    return this.props.attachmentId;
+  }
+
+  static create(
+    props: AnswerAttachmentProps,
+    id?: UniqueEntityID
+  ): AnswerAttachment {
+    return new AnswerAttachment(props, id);
+  }
+}
