--- src/domain/forum/enterprise/entities/answer-attatchment.ts
@@ -1,24 +1,0 @@
-import { Entity } from "@/core/entities/entity";
-import { UniqueEntityID } from "@/core/entities/unique-entity-id";
-
-interface QuestionAttachmentProps {
-  questionId: UniqueEntityID;
-  attachmentId: UniqueEntityID;
-}
-
-export class QuestionAttachment extends Entity<QuestionAttachmentProps> {
-  get questionId() {
-    return this.props.questionId;
-  }
-
-  get attachmentId() {
-    return this.props.attachmentId;
-  }
-
-  static create(
-    props: QuestionAttachmentProps,
-    id?: UniqueEntityID
-  ): QuestionAttachment {
-    return new QuestionAttachment(props, id);
-  }
-}
