+++ src/domain/forum/enterprise/entities/question-attachment-list.ts
@@ -1,5 +1,5 @@
 import { WatchedList } from "@/core/entities/watched-list";
-import { QuestionAttachment } from "./answer-attatchment";
+import { QuestionAttachment } from "./answer-attachment";
 
 export class QuestionAttachmentList extends WatchedList<QuestionAttachment> {
   compareItems(a: QuestionAttachment, b: QuestionAttachment): boolean {
